-create a ClientRequestComposerProvider to provide composer based on composer name. the consumers can register by their customized composers
