+++ client/src/main/java/com/networknt/client/oauth/IClientRequestComposable.java
@@ -2,7 +2,7 @@ package com.networknt.client.oauth;
 
 import io.undertow.client.ClientRequest;
 
-public interface ClientRequestComposable {
+public interface IClientRequestComposable {
     ClientRequest ComposeClientRequest(TokenRequest tokenRequest);
     String ComposeRequestBody(TokenRequest tokenRequest);
 }
